+++ src/main/java/fr/gamejam/papee/engine/game/GGame.java
@@ -1,16 +1,40 @@
 package fr.gamejam.papee.engine.game;
 
+import fr.gamejam.papee.engine.components.GWindow;
+import fr.gamejam.papee.engine.fx.GParticle;
+import fr.gamejam.papee.engine.objects.GObject;
+
+import java.util.ArrayList;
+
 public abstract class GGame {
+
+    public static ArrayList<GObject> objects = new ArrayList<GObject>();
+    public static GWindow window;
 
     public GGame() {}
 
     /**
      * Called every frame (updating position ...)
      */
-    public abstract void update();
+    public void update() {
+        for (int i = 0; i < objects.size(); i++) {
+            GObject o = objects.get(i);
+            o.update();
+            if (o instanceof GParticle) {
+                if (((GParticle) o).getLifetime() < 0) {
+                    objects.remove(o);
+                }
+            }
+        }
+    }
 
     /**
      * Called every frame (rendering objects ...)
      */
-    public abstract void render();
+    public void render() {
+        for (int i = 0; i < objects.size(); i++) {
+            GObject o = objects.get(i);
+            o.render();
+        }
+    }
 }
